Refactored Sellar case for test of FASTOADProblemConfigurator
