+++ src/fastoad/io/configuration/tests/data/conf_sellar_example/disc2.py
@@ -12,14 +12,22 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import openmdao.api as om
+
 from fastoad.module_management.constants import ModelDomain
 from fastoad.module_management.service_registry import RegisterOpenMDAOSystem
-from .disc2_base import Disc2Base
 
 
 @RegisterOpenMDAOSystem("configuration_test.sellar.disc2", domain=ModelDomain.GEOMETRY)
-class Disc2(Disc2Base):
+class Disc2(om.ExplicitComponent):
     """ An OpenMDAO component to encapsulate Disc2 discipline """
+
+    def setup(self):
+        self.add_input("z", val=[5, 2], desc="", units="m**2")  # for testing non-None units
+        self.add_input("y1", val=1.0, desc="")
+
+        self.add_output("y2", val=1.0, desc="")
+        self.declare_partials("*", "*", method="fd")
 
     # pylint: disable=invalid-name
     def compute(self, inputs, outputs, discrete_inputs=None, discrete_outputs=None):
